Z3.Lang.Pow2

Description

Documentation

declarePow2 :: IsInt a => Z3 (Expr a -> Expr a)

Raise to the power of 2. Axiomatization of the 2^n function. Most likely Z3 is going to diverge if you use 2^n to specify a satisfiable problem, since it cannot construct a recursive definition for 2^n, but it should work fine for unsatisfiable instances.