Z3.Monad

Description

A simple monadic wrapper for Base.

Z3 monad

class (Monad m, MonadIO m) => MonadZ3 m where

Methods

getSolver :: m (Maybe Solver)

getContext :: m Context

data Z3 a

module Z3.Opts

data Logic

Warning: New Z3 API support is still incomplete and fragile: you may experience segmentation faults!

Solvers available in Z3.

These are described at http://smtlib.cs.uiowa.edu/logics.html

WARNING: Support for solvers is fragile, you may experience segmentation faults!

Constructors

AUFLIA
Closed formulas over the theory of linear integer arithmetic and arrays extended with free sort and function symbols but restricted to arrays with integer indices and values.

AUFLIRA
Closed linear formulas with free sort and function symbols over one- and two-dimentional arrays of integer index and real value.

AUFNIRA
Closed formulas with free function and predicate symbols over a theory of arrays of arrays of integer index and real value.

LRA
Closed linear formulas in linear real arithmetic.

QF_ABV
Closed quantifier-free formulas over the theory of bitvectors and bitvector arrays.

QF_AUFBV
Closed quantifier-free formulas over the theory of bitvectors and bitvector arrays extended with free sort and function symbols.

QF_AUFLIA
Closed quantifier-free linear formulas over the theory of integer arrays extended with free sort and function symbols.

QF_AX
Closed quantifier-free formulas over the theory of arrays with extensionality.

QF_BV
Closed quantifier-free formulas over the theory of fixed-size bitvectors.

QF_IDL
Difference Logic over the integers. In essence, Boolean combinations of inequations of the form x - y < b where x and y are integer variables and b is an integer constant.

QF_LIA
Unquantified linear integer arithmetic. In essence, Boolean combinations of inequations between linear polynomials over integer variables.

QF_LRA
Unquantified linear real arithmetic. In essence, Boolean combinations of inequations between linear polynomials over real variables.

QF_NIA
Quantifier-free integer arithmetic.

QF_NRA
Quantifier-free real arithmetic.

QF_RDL
Difference Logic over the reals. In essence, Boolean combinations of inequations of the form x - y < b where x and y are real variables and b is a rational constant.

QF_UF
Unquantified formulas built over a signature of uninterpreted (i.e., free) sort and function symbols.

QF_UFBV
Unquantified formulas over bitvectors with uninterpreted sort function and symbols.

QF_UFIDL
Difference Logic over the integers (in essence) but with uninterpreted sort and function symbols.

QF_UFLIA
Unquantified linear integer arithmetic with uninterpreted sort and function symbols.

QF_UFLRA
Unquantified linear real arithmetic with uninterpreted sort and function symbols.

QF_UFNRA
Unquantified non-linear real arithmetic with uninterpreted sort and function symbols.

UFLRA
Linear real arithmetic with uninterpreted sort and function symbols.

UFNIA
Non-linear integer arithmetic with uninterpreted sort and function symbols.

evalZ3 :: Z3 a -> IO a

Eval a Z3 script with default configuration options.

evalZ3With :: Maybe Logic -> Opts -> Z3 a -> IO a

Eval a Z3 script.

Z3 enviroments

data Z3Env

Z3 environment.

newEnv :: Maybe Logic -> Opts -> IO Z3Env

Create a new Z3 environment.

Until we move to Z3 API 4.0 you need to manually freed this environment using delEnv.

delEnv :: Z3Env -> IO ()

Free a Z3 environment.

evalZ3WithEnv :: Z3 a -> Z3Env -> IO a

Eval a Z3 script with a given environment.

Environments may facilitate running many queries under the same logical context.

Note that an environment may change after each query. If you want to preserve the same environment then use local, as in evalZ3WithEnv env (local query).

Types

data Symbol

A Z3 symbol.

Used to name types, constants and functions.

data AST

A Z3 AST node.

This is the data-structure used in Z3 to represent terms, formulas and types.

data Sort

A kind of AST representing types.

data FuncDecl

A kind of AST representing function symbols.

data App

A kind of AST representing constant and function declarations.

data Pattern

A kind of AST representing pattern and multi-patterns to guide quantifier instantiation.

data Model

A model for the constraints asserted into the logical context.

data Context

A Z3 logical context.

data FuncInterp

An interpretation of a function in a model.

data FuncEntry

Representation of the value of a Z3_func_interp at a particular point.

data FuncModel

The interpretation of a function.

Constructors

FuncModel

Fields

interpMap :: [([AST], AST)]

Mapping from arguments to values.

interpElse :: AST

Default value.

data Solver

A Z3 solver engine.

A(n) (incremental) solver, possibly specialized by a particular tactic or logic.

Satisfiability result

data Result

Result of a satisfiability check.

This corresponds to the z3_lbool type in the C API.

Constructors

Sat

Unsat

Undef

Context

contextToString :: MonadZ3 z3 => z3 String

Convert Z3's logical context into a string.

showContext :: MonadZ3 z3 => z3 String

Alias for contextToString.

Symbols

mkIntSymbol :: (MonadZ3 z3, Integral i) => i -> z3 Symbol

Create a Z3 symbol using an integer.

mkStringSymbol :: MonadZ3 z3 => String -> z3 Symbol

Create a Z3 symbol using a string.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafebb0d3c212927cf7834c3a20a84ecae

Sorts

mkUninterpretedSort :: MonadZ3 z3 => Symbol -> z3 Sort

Create a free (uninterpreted) type using the given name (symbol).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga736e88741af1c178cbebf94c49aa42de

mkBoolSort :: MonadZ3 z3 => z3 Sort

Create the boolean type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacdc73510b69a010b71793d429015f342

mkIntSort :: MonadZ3 z3 => z3 Sort

Create the integer type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6cd426ab5748653b77d389fd3eac1015

mkRealSort :: MonadZ3 z3 => z3 Sort

Create the real type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga40ef93b9738485caed6dc84631c3c1a0

mkBvSort :: MonadZ3 z3 => Int -> z3 Sort

Create a bit-vector type of the given size.

This type can also be seen as a machine integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaeed000a1bbb84b6ca6fdaac6cf0c1688

mkArraySort :: MonadZ3 z3 => Sort -> Sort -> z3 Sort

Create an array type

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafe617994cce1b516f46128e448c84445

mkTupleSort

Arguments

:: MonadZ3 z3

=> Symbol
Name of the sort

-> [(Symbol, Sort)]
Name and sort of each field

-> z3 (Sort, FuncDecl, [FuncDecl])
Resulting sort, and function declarations for the constructor and projections.

Create a tuple type

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7156b9c0a76a28fae46c81f8e3cdf0f1

Constants and Applications

mkFuncDecl :: MonadZ3 z3 => Symbol -> [Sort] -> Sort -> z3 FuncDecl

A Z3 function

mkApp :: MonadZ3 z3 => FuncDecl -> [AST] -> z3 AST

Create a constant or function application.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga33a202d86bf628bfab9b6f437536cebe

mkConst :: MonadZ3 z3 => Symbol -> Sort -> z3 AST

Declare and create a constant.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga093c9703393f33ae282ec5e8729354ef

mkTrue :: MonadZ3 z3 => z3 AST

Create an AST node representing true.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae898e7380409bbc57b56cc5205ef1db7

mkFalse :: MonadZ3 z3 => z3 AST

Create an AST node representing false.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5952ac17671117a02001fed6575c778d

mkEq :: MonadZ3 z3 => AST -> AST -> z3 AST

Create an AST node representing l = r.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga95a19ce675b70e22bb0401f7137af37c

mkNot :: MonadZ3 z3 => AST -> z3 AST

Create an AST node representing not(a).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga3329538091996eb7b3dc677760a61072

mkIte :: MonadZ3 z3 => AST -> AST -> AST -> z3 AST

Create an AST node representing an if-then-else: ite(t1, t2, t3).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga94417eed5c36e1ad48bcfc8ad6e83547

mkIff :: MonadZ3 z3 => AST -> AST -> z3 AST

Create an AST node representing t1 iff t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga930a8e844d345fbebc498ac43a696042

mkImplies :: MonadZ3 z3 => AST -> AST -> z3 AST

Create an AST node representing t1 implies t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac829c0e25bbbd30343bf073f7b524517

mkXor :: MonadZ3 z3 => AST -> AST -> z3 AST

Create an AST node representing t1 xor t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacc6d1b848032dec0c4617b594d4229ec

mkAnd :: MonadZ3 z3 => [AST] -> z3 AST

Create an AST node representing args[0] and ... and args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacde98ce4a8ed1dde50b9669db4838c61

mkOr :: MonadZ3 z3 => [AST] -> z3 AST

Create an AST node representing args[0] or ... or args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga00866d16331d505620a6c515302021f9

mkDistinct :: MonadZ3 z3 => [AST] -> z3 AST

The distinct construct is used for declaring the arguments pairwise distinct.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaa076d3a668e0ec97d61744403153ecf7

mkAdd :: MonadZ3 z3 => [AST] -> z3 AST

Create an AST node representing args[0] + ... + args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4e4ac0a4e53eee0b4b0ef159ed7d0cd5

mkMul :: MonadZ3 z3 => [AST] -> z3 AST

Create an AST node representing args[0] * ... * args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab9affbf8401a18eea474b59ad4adc890

mkSub :: MonadZ3 z3 => [AST] -> z3 AST

Create an AST node representing args[0] - ... - args[num_args - 1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4f5fea9b683f9e674fd8f14d676cc9a9

mkUnaryMinus :: MonadZ3 z3 => AST -> z3 AST

Create an AST node representing -arg.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gadcd2929ad732937e25f34277ce4988ea

mkDiv :: MonadZ3 z3 => AST -> AST -> z3 AST

Create an AST node representing arg1 div arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1ac60ee8307af8d0b900375914194ff3

mkMod :: MonadZ3 z3 => AST -> AST -> z3 AST

Create an AST node representing arg1 mod arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga8e350ac77e6b8fe805f57efe196e7713

mkRem :: MonadZ3 z3 => AST -> AST -> z3 AST

Create an AST node representing arg1 rem arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga2fcdb17f9039bbdaddf8a30d037bd9ff

mkLt :: MonadZ3 z3 => AST -> AST -> z3 AST

Create less than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga58a3dc67c5de52cf599c346803ba1534

mkLe :: MonadZ3 z3 => AST -> AST -> z3 AST

Create less than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaa9a33d11096841f4e8c407f1578bc0bf

mkGt :: MonadZ3 z3 => AST -> AST -> z3 AST

Create greater than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga46167b86067586bb742c0557d7babfd3

mkGe :: MonadZ3 z3 => AST -> AST -> z3 AST

Create greater than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad9245cbadb80b192323d01a8360fb942

mkInt2Real :: MonadZ3 z3 => AST -> z3 AST

Coerce an integer to a real.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7130641e614c7ebafd28ae16a7681a21

mkReal2Int :: MonadZ3 z3 => AST -> z3 AST

Coerce a real to an integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga759b6563ba1204aae55289009a3fdc6d

mkIsInt :: MonadZ3 z3 => AST -> z3 AST

Check if a real number is an integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaac2ad0fb04e4900fdb4add438d137ad3

Bit-vectors

mkBvnot :: MonadZ3 z3 => AST -> z3 AST

Bitwise negation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga36cf75c92c54c1ca633a230344f23080

mkBvredand :: MonadZ3 z3 => AST -> z3 AST

Take conjunction of bits in vector, return vector of length 1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaccc04f2b58903279b1b3be589b00a7d8

mkBvredor :: MonadZ3 z3 => AST -> z3 AST

Take disjunction of bits in vector, return vector of length 1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafd18e127c0586abf47ad9cd96895f7d2

mkBvand :: MonadZ3 z3 => AST -> AST -> z3 AST

Bitwise and.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab96e0ea55334cbcd5a0e79323b57615d

mkBvor :: MonadZ3 z3 => AST -> AST -> z3 AST

Bitwise or.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga77a6ae233fb3371d187c6d559b2843f5

mkBvxor :: MonadZ3 z3 => AST -> AST -> z3 AST

Bitwise exclusive-or.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga0a3821ea00b1c762205f73e4bc29e7d8

mkBvnand :: MonadZ3 z3 => AST -> AST -> z3 AST

Bitwise nand.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga96dc37d36efd658fff5b2b4df49b0e61

mkBvnor :: MonadZ3 z3 => AST -> AST -> z3 AST

Bitwise nor.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gabf15059e9e8a2eafe4929fdfd259aadb

mkBvxnor :: MonadZ3 z3 => AST -> AST -> z3 AST

Bitwise xnor.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga784f5ca36a4b03b93c67242cc94b21d6

mkBvneg :: MonadZ3 z3 => AST -> z3 AST

Standard two's complement unary minus.

Reference: <http:research.microsoft.comen-usumredmondprojectsz3group__capi.html#ga0c78be00c03eda4ed6a983224ed5c7b7

mkBvadd :: MonadZ3 z3 => AST -> AST -> z3 AST

Standard two's complement addition.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga819814e33573f3f9948b32fdc5311158

mkBvsub :: MonadZ3 z3 => AST -> AST -> z3 AST

Standard two's complement subtraction.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga688c9aa1347888c7a51be4e46c19178e

mkBvmul :: MonadZ3 z3 => AST -> AST -> z3 AST

Standard two's complement multiplication.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6abd3dde2a1ceff1704cf7221a72258c

mkBvudiv :: MonadZ3 z3 => AST -> AST -> z3 AST

Unsigned division.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga56ce0cd61666c6f8cf5777286f590544

mkBvsdiv :: MonadZ3 z3 => AST -> AST -> z3 AST

Two's complement signed division.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad240fedb2fda1c1005b8e9d3c7f3d5a0

mkBvurem :: MonadZ3 z3 => AST -> AST -> z3 AST

Unsigned remainder.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5df4298ec835e43ddc9e3e0bae690c8d

mkBvsrem :: MonadZ3 z3 => AST -> AST -> z3 AST

Two's complement signed remainder (sign follows dividend).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga46c18a3042fca174fe659d3185693db1

mkBvsmod :: MonadZ3 z3 => AST -> AST -> z3 AST

Two's complement signed remainder (sign follows divisor).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga95dac8e6eecb50f63cb82038560e0879

mkBvult :: MonadZ3 z3 => AST -> AST -> z3 AST

Unsigned less than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5774b22e93abcaf9b594672af6c7c3c4

mkBvslt :: MonadZ3 z3 => AST -> AST -> z3 AST

Two's complement signed less than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga8ce08af4ed1fbdf08d4d6e63d171663a

mkBvule :: MonadZ3 z3 => AST -> AST -> z3 AST

Unsigned less than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab738b89de0410e70c089d3ac9e696e87

mkBvsle :: MonadZ3 z3 => AST -> AST -> z3 AST

Two's complement signed less than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab7c026feb93e7d2eab180e96f1e6255d

mkBvuge :: MonadZ3 z3 => AST -> AST -> z3 AST

Unsigned greater than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gade58fbfcf61b67bf8c4a441490d3c4df

mkBvsge :: MonadZ3 z3 => AST -> AST -> z3 AST

Two's complement signed greater than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaeec3414c0e8a90a6aa5a23af36bf6dc5

mkBvugt :: MonadZ3 z3 => AST -> AST -> z3 AST

Unsigned greater than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga063ab9f16246c99e5c1c893613927ee3

mkBvsgt :: MonadZ3 z3 => AST -> AST -> z3 AST

Two's complement signed greater than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4e93a985aa2a7812c7c11a2c65d7c5f0

mkConcat :: MonadZ3 z3 => AST -> AST -> z3 AST

Concatenate the given bit-vectors.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae774128fa5e9ff7458a36bd10e6ca0fa

mkExtract :: MonadZ3 z3 => Int -> Int -> AST -> z3 AST

Extract the bits high down to low from a bitvector of size m to yield a new bitvector of size n, where n = high - low + 1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga32d2fe7563f3e6b114c1b97b205d4317

mkSignExt :: MonadZ3 z3 => Int -> AST -> z3 AST

Sign-extend of the given bit-vector to the (signed) equivalent bitvector of size m+i, where m is the size of the given bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad29099270b36d0680bb54b560353c10e

mkZeroExt :: MonadZ3 z3 => Int -> AST -> z3 AST

Extend the given bit-vector with zeros to the (unsigned) equivalent bitvector of size m+i, where m is the size of the given bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac9322fae11365a78640baf9078c428b3

mkRepeat :: MonadZ3 z3 => Int -> AST -> z3 AST

Repeat the given bit-vector up length i.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga03e81721502ea225c264d1f556c9119d

mkBvshl :: MonadZ3 z3 => AST -> AST -> z3 AST

Shift left.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac8d5e776c786c1172fa0d7dfede454e1

mkBvlshr :: MonadZ3 z3 => AST -> AST -> z3 AST

Logical shift right.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac59645a6edadad79a201f417e4e0c512

mkBvashr :: MonadZ3 z3 => AST -> AST -> z3 AST

Arithmetic shift right.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga674b580ad605ba1c2c9f9d3748be87c4

mkRotateLeft :: MonadZ3 z3 => Int -> AST -> z3 AST

Rotate bits of t1 to the left i times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4932b7d08fea079dd903cd857a52dcda

mkRotateRight :: MonadZ3 z3 => Int -> AST -> z3 AST

Rotate bits of t1 to the right i times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga3b94e1bf87ecd1a1858af8ebc1da4a1c

mkExtRotateLeft :: MonadZ3 z3 => AST -> AST -> z3 AST

Rotate bits of t1 to the left t2 times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaf46f1cb80e5a56044591a76e7c89e5e7

mkExtRotateRight :: MonadZ3 z3 => AST -> AST -> z3 AST

Rotate bits of t1 to the right t2 times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gabb227526c592b523879083f12aab281f

mkInt2bv :: MonadZ3 z3 => Int -> AST -> z3 AST

Create an n bit bit-vector from the integer argument t1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga35f89eb05df43fbd9cce7200cc1f30b5

mkBv2int :: MonadZ3 z3 => AST -> Bool -> z3 AST

Create an integer from the bit-vector argument t1. If is_signed is false, then the bit-vector t1 is treated as unsigned. So the result is non-negative and in the range [0..2^N-1], where N are the number of bits in t1. If is_signed is true, t1 is treated as a signed bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac87b227dc3821d57258d7f53a28323d4

mkBvnegNoOverflow :: MonadZ3 z3 => AST -> z3 AST

Check that bit-wise negation does not overflow when t1 is interpreted as a signed bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae9c5d72605ddcd0e76657341eaccb6c7

mkBvaddNoOverflow :: MonadZ3 z3 => AST -> AST -> Bool -> z3 AST

Create a predicate that checks that the bit-wise addition of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga88f6b5ec876f05e0d7ba51e96c4b077f

mkBvaddNoUnderflow :: MonadZ3 z3 => AST -> AST -> z3 AST

Create a predicate that checks that the bit-wise signed addition of t1 and t2 does not underflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1e2b1927cf4e50000c1600d47a152947

mkBvsubNoOverflow :: MonadZ3 z3 => AST -> AST -> z3 AST

Create a predicate that checks that the bit-wise signed subtraction of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga785f8127b87e0b42130e6d8f52167d7c

mkBvsubNoUnderflow :: MonadZ3 z3 => AST -> AST -> z3 AST

Create a predicate that checks that the bit-wise subtraction of t1 and t2 does not underflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6480850f9fa01e14aea936c88ff184c4

mkBvmulNoOverflow :: MonadZ3 z3 => AST -> AST -> Bool -> z3 AST

Create a predicate that checks that the bit-wise multiplication of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga86f4415719d295a2f6845c70b3aaa1df

mkBvmulNoUnderflow :: MonadZ3 z3 => AST -> AST -> z3 AST

Create a predicate that checks that the bit-wise signed multiplication of t1 and t2 does not underflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga501ccc01d737aad3ede5699741717fda

mkBvsdivNoOverflow :: MonadZ3 z3 => AST -> AST -> z3 AST

Create a predicate that checks that the bit-wise signed division of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaa17e7b2c33dfe2abbd74d390927ae83e

Arrays

mkSelect :: MonadZ3 z3 => AST -> AST -> z3 AST

Array read. The argument a is the array and i is the index of the array that gets read.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga38f423f3683379e7f597a7fe59eccb67

mkStore :: MonadZ3 z3 => AST -> AST -> AST -> z3 AST

Array update.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae305a4f54b4a64f7e5973ae6ccb13593

mkConstArray :: MonadZ3 z3 => Sort -> AST -> z3 AST

Create the constant array.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga84ea6f0c32b99c70033feaa8f00e8f2d

mkMap :: MonadZ3 z3 => FuncDecl -> [AST] -> z3 AST

map f on the the argument arrays.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga9150242d9430a8c3d55d2ca3b9a4362d

mkArrayDefault :: MonadZ3 z3 => AST -> z3 AST

Access the array default value. Produces the default range value, for arrays that can be represented as finite maps with a default range value.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga78e89cca82f0ab4d5f4e662e5e5fba7d

Numerals

mkNumeral :: MonadZ3 z3 => String -> Sort -> z3 AST

Create a numeral of a given sort.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac8aca397e32ca33618d8024bff32948c

mkInt :: (MonadZ3 z3, Integral a) => a -> z3 AST

Create a numeral of sort int.

mkReal :: (MonadZ3 z3, Real r) => r -> z3 AST

Create a numeral of sort real.

Quantifiers

mkPattern :: MonadZ3 z3 => [AST] -> z3 Pattern

mkBound :: MonadZ3 z3 => Int -> Sort -> z3 AST

mkForall :: MonadZ3 z3 => [Pattern] -> [Symbol] -> [Sort] -> AST -> z3 AST

mkExists :: MonadZ3 z3 => [Pattern] -> [Symbol] -> [Sort] -> AST -> z3 AST

mkForallConst :: MonadZ3 z3 => [Pattern] -> [App] -> AST -> z3 AST

mkExistsConst :: MonadZ3 z3 => [Pattern] -> [App] -> AST -> z3 AST

Accessors

getBvSortSize :: MonadZ3 z3 => Sort -> z3 Int

Return the size of the given bit-vector sort.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga8fc3550edace7bc046e16d1f96ddb419

getBool :: MonadZ3 z3 => AST -> z3 (Maybe Bool)

Returns Just True, Just False, or Nothing for undefined.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga133aaa1ec31af9b570ed7627a3c8c5a4

getInt :: MonadZ3 z3 => AST -> z3 Integer

Return the integer value

getReal :: MonadZ3 z3 => AST -> z3 Rational

Return rational value

toApp :: MonadZ3 z3 => AST -> z3 App

Cast AST into an App.

Models

eval :: MonadZ3 z3 => Model -> AST -> z3 (Maybe AST)

Evaluate an AST node in the given model.

evalT :: (MonadZ3 z3, Traversable t) => Model -> t AST -> z3 (Maybe (t AST))

Evaluate a collection of AST nodes in the given model.

evalFunc :: MonadZ3 z3 => Model -> FuncDecl -> z3 (Maybe FuncModel)

Get function as a list of argument/value pairs.

evalArray :: MonadZ3 z3 => Model -> AST -> z3 (Maybe FuncModel)

Get array as a list of argument/value pairs, if it is represented as a function (ie, using as-array).

getFuncInterp :: MonadZ3 z3 => Model -> FuncDecl -> z3 (Maybe FuncInterp)

Return the interpretation of the function f in the model m. Return NULL, if the model does not assign an interpretation for f. That should be interpreted as: the f does not matter.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafb9cc5eca9564d8a849c154c5a4a8633

isAsArray :: MonadZ3 z3 => AST -> z3 Bool

The (_ as-array f) AST node is a construct for assigning interpretations for arrays in Z3. It is the array such that forall indices i we have that (select (_ as-array f) i) is equal to (f i). This procedure returns Z3_TRUE if the a is an as-array AST node.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4674da67d226bfb16861829b9f129cfa

getAsArrayFuncDecl :: MonadZ3 z3 => AST -> z3 FuncDecl

Return the function declaration f associated with a (_ as_array f) node.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7d9262dc6e79f2aeb23fd4a383589dda

funcInterpGetNumEntries :: MonadZ3 z3 => FuncInterp -> z3 Int

Return the number of entries in the given function interpretation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga2bab9ae1444940e7593729beec279844

funcInterpGetEntry :: MonadZ3 z3 => FuncInterp -> Int -> z3 FuncEntry

Return a point of the given function intepretation. It represents the value of f in a particular point.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaf157e1e1cd8c0cfe6a21be6370f659da

funcInterpGetElse :: MonadZ3 z3 => FuncInterp -> z3 AST

Return the 'else' value of the given function interpretation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga46de7559826ba71b8488d727cba1fb64

funcInterpGetArity :: MonadZ3 z3 => FuncInterp -> z3 Int

Return the arity (number of arguments) of the given function interpretation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaca22cbdb6f7787aaae5d814f2ab383d8

funcEntryGetValue :: MonadZ3 z3 => FuncEntry -> z3 AST

Return the value of this point.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga9fd65e2ab039aa8e40608c2ecf7084da

funcEntryGetNumArgs :: MonadZ3 z3 => FuncEntry -> z3 Int

Return the number of arguments in a Z3_func_entry object.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga51aed8c5bc4b1f53f0c371312de3ce1a

funcEntryGetArg :: MonadZ3 z3 => FuncEntry -> Int -> z3 AST

Return an argument of a Z3_func_entry object.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6fe03fe3c824fceb52766a4d8c2cbeab

modelToString :: MonadZ3 z3 => Model -> z3 String

Convert the given model into a string.

showModel :: MonadZ3 z3 => Model -> z3 String

Alias for modelToString.

Constraints

assertCnstr :: MonadZ3 z3 => AST -> z3 ()

Assert a constraing into the logical context.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1a05ff73a564ae7256a2257048a4680a

check :: MonadZ3 z3 => z3 Result

Check whether the given logical context is consistent or not.

getModel :: MonadZ3 z3 => z3 (Result, Maybe Model)

Get model.

Reference : http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaff310fef80ac8a82d0a51417e073ec0a

delModel :: MonadZ3 z3 => Model -> z3 ()

Delete a model object.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga0cc98d3ce68047f873e119bccaabdbee

withModel :: (Applicative z3, MonadZ3 z3) => (Model -> z3 a) -> z3 (Result, Maybe a)

push :: MonadZ3 z3 => z3 ()

Create a backtracking point.

pop :: MonadZ3 z3 => Int -> z3 ()

Backtrack n backtracking points.

local :: MonadZ3 z3 => z3 a -> z3 a

Run a query and restore the initial logical context.

This is a shorthand for push, run the query, and pop.

reset :: MonadZ3 z3 => z3 ()

Backtrack all the way.

getNumScopes :: MonadZ3 z3 => z3 Int

Get number of backtracking points.

String Conversion

data ASTPrintMode

Pretty-printing mode for converting ASTs to strings. The mode can be one of the following:

Z3_PRINT_SMTLIB_FULL: Print AST nodes in SMTLIB verbose format.

Z3_PRINT_LOW_LEVEL: Print AST nodes using a low-level format.

Z3_PRINT_SMTLIB_COMPLIANT: Print AST nodes in SMTLIB 1.x compliant format.

Z3_PRINT_SMTLIB2_COMPLIANT: Print AST nodes in SMTLIB 2.x compliant format.

Constructors

Z3_PRINT_SMTLIB_FULL

Z3_PRINT_LOW_LEVEL

Z3_PRINT_SMTLIB_COMPLIANT

Z3_PRINT_SMTLIB2_COMPLIANT

setASTPrintMode :: MonadZ3 z3 => ASTPrintMode -> z3 ()

Set the mode for converting expressions to strings.

astToString :: MonadZ3 z3 => AST -> z3 String

Convert an AST to a string.

patternToString :: MonadZ3 z3 => Pattern -> z3 String

Convert a pattern to a string.

sortToString :: MonadZ3 z3 => Sort -> z3 String

Convert a sort to a string.

funcDeclToString :: MonadZ3 z3 => FuncDecl -> z3 String

Convert a FuncDecl to a string.

solverToString :: MonadZ3 z3 => z3 String

Convert the solver to a string.

benchmarkToSMTLibString

Arguments

:: MonadZ3 z3

=> String
name

-> String
logic

-> String
status

-> String
attributes

-> [AST]
assumptions1

-> AST
formula

-> z3 String

Convert the given benchmark into SMT-LIB formatted string.

The output format can be configured via setASTPrintMode.

Miscellaneous

data Version

Constructors

Version

Fields

z3Major :: !Int

z3Minor :: !Int

z3Build :: !Int

z3Revision :: !Int

getVersion :: MonadZ3 z3 => z3 Version

Return Z3 version number information.