Z3.Opts

Description

Configuring Z3.

Z3 has plenty of configuration options and these vary quite a lot across Z3 versions, being hard to design a proper abstraction. We decided to keep this simple.

Z3 configuration

data Opts

Z3 configuration.

setOpts :: Config -> Opts -> IO ()

Set configuration.

If you are using Lang or Monad interfaces, you don't need to call this function directly, just pass your Opts to evalZ3*.

stdOpts :: Opts

Default configuration.

(+?) :: Opts -> Opts -> Opts

Append configurations.

Z3 options

opt :: OptValue val => String -> val -> Opts

Set a configuration option.

class OptValue val

Values for Z3 options.