Z3.Base

Description

Low-level bindings to Z3 API.

Types

data Config

A Z3 configuration object.

data Context

A Z3 logical context.

data Symbol

A Z3 Lisp-link symbol.

data AST

A Z3 AST node.

data Sort

Kind of Z3 AST representing types.

data FuncDecl

Kind of AST used to represent function symbols.

data App

A kind of Z3 AST used to represent constant and function declarations.

data Pattern

A kind of AST used to represent pattern and multi-patterns used to guide quantifier instantiation.

data Model

A model for the constraints asserted into the logical context.

data FuncInterp

A interpretation of a function.

data FuncEntry

data Params

A Z3 parameter set. Starting at Z3 4.0, parameter sets are used to configure many components such as: simplifiers, tactics, solvers, etc.

data Solver

A Z3 solver engine

Satisfiability result

data Result

Result of a satisfiability check.

Constructors

Sat

Unsat

Undef

Configuration

mkConfig :: IO Config

Create a configuration.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7d6c40d9b79fe8a8851cc8540970787f

delConfig :: Config -> IO ()

Delete a configuration.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5e620acf5d55d0271097c9bb97219774

withConfig :: (Config -> IO a) -> IO a

Run a computation using a temporally created configuration.

setParamValue :: Config -> String -> String -> IO ()

Set a configuration parameter.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga001ade87a1671fe77d7e53ed0f4f1ec3

See: http://research.microsoft.com/en-us/um/redmond/projects/z3/config.html

Context

mkContext :: Config -> IO Context

Create a context using the given configuration.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga0bd93cfab4d749dd3e2f2a4416820a46

delContext :: Context -> IO ()

Delete the given logical context.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga556eae80ed43ab13e1e7dc3b38c35200

withContext :: Config -> (Context -> IO a) -> IO a

Run a computation using a temporally created context.

contextToString :: Context -> IO String

Convert the given logical context into a string.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga165e38ddfc928f586cb738cdf6c5f216

showContext :: Context -> IO String

Alias for contextToString.

Symbols

mkIntSymbol :: Integral int => Context -> int -> IO Symbol

Create a Z3 symbol using an integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga3df806baf6124df3e63a58cf23e12411

mkStringSymbol :: Context -> String -> IO Symbol

Create a Z3 symbol using a string.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafebb0d3c212927cf7834c3a20a84ecae

Sorts

mkUninterpretedSort :: Context -> Symbol -> IO Sort

Create a free (uninterpreted) type using the given name (symbol).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga736e88741af1c178cbebf94c49aa42de

mkBoolSort :: Context -> IO Sort

Create the Boolean type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacdc73510b69a010b71793d429015f342

mkIntSort :: Context -> IO Sort

Create an integer type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6cd426ab5748653b77d389fd3eac1015

mkRealSort :: Context -> IO Sort

Create a real type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga40ef93b9738485caed6dc84631c3c1a0

mkBvSort :: Context -> Int -> IO Sort

Create a bit-vector type of the given size.

This type can also be seen as a machine integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaeed000a1bbb84b6ca6fdaac6cf0c1688

mkArraySort :: Context -> Sort -> Sort -> IO Sort

Create an array type

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafe617994cce1b516f46128e448c84445

mkTupleSort

Arguments

:: Context
Context

-> Symbol
Name of the sort

-> [(Symbol, Sort)]
Name and sort of each field

-> IO (Sort, FuncDecl, [FuncDecl])
Resulting sort, and function declarations for the constructor and projections.

Create a tuple type

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7156b9c0a76a28fae46c81f8e3cdf0f1

Constants and Applications

mkFuncDecl :: Context -> Symbol -> [Sort] -> Sort -> IO FuncDecl

A Z3 function

mkApp :: Context -> FuncDecl -> [AST] -> IO AST

Create a constant or function application.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga33a202d86bf628bfab9b6f437536cebe

mkConst :: Context -> Symbol -> Sort -> IO AST

Declare and create a constant.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga093c9703393f33ae282ec5e8729354ef

mkTrue :: Context -> IO AST

Create an AST node representing true.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae898e7380409bbc57b56cc5205ef1db7

mkFalse :: Context -> IO AST

Create an AST node representing false.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5952ac17671117a02001fed6575c778d

mkEq :: Context -> AST -> AST -> IO AST

Create an AST node representing l = r.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga95a19ce675b70e22bb0401f7137af37c

mkNot :: Context -> AST -> IO AST

Create an AST node representing not(a).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga3329538091996eb7b3dc677760a61072

mkIte :: Context -> AST -> AST -> AST -> IO AST

Create an AST node representing an if-then-else: ite(t1, t2, t3).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga94417eed5c36e1ad48bcfc8ad6e83547

mkIff :: Context -> AST -> AST -> IO AST

Create an AST node representing t1 iff t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga930a8e844d345fbebc498ac43a696042

mkImplies :: Context -> AST -> AST -> IO AST

Create an AST node representing t1 implies t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac829c0e25bbbd30343bf073f7b524517

mkXor :: Context -> AST -> AST -> IO AST

Create an AST node representing t1 xor t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacc6d1b848032dec0c4617b594d4229ec

mkAnd :: Context -> [AST] -> IO AST

Create an AST node representing args[0] and ... and args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacde98ce4a8ed1dde50b9669db4838c61

mkOr :: Context -> [AST] -> IO AST

Create an AST node representing args[0] or ... or args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga00866d16331d505620a6c515302021f9

mkDistinct :: Context -> [AST] -> IO AST

The distinct construct is used for declaring the arguments pairwise distinct.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaa076d3a668e0ec97d61744403153ecf7

mkAdd :: Context -> [AST] -> IO AST

Create an AST node representing args[0] + ... + args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4e4ac0a4e53eee0b4b0ef159ed7d0cd5

mkMul :: Context -> [AST] -> IO AST

Create an AST node representing args[0] * ... * args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab9affbf8401a18eea474b59ad4adc890

mkSub :: Context -> [AST] -> IO AST

Create an AST node representing args[0] - ... - args[num_args - 1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4f5fea9b683f9e674fd8f14d676cc9a9

mkUnaryMinus :: Context -> AST -> IO AST

Create an AST node representing -arg.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gadcd2929ad732937e25f34277ce4988ea

mkDiv :: Context -> AST -> AST -> IO AST

Create an AST node representing arg1 div arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1ac60ee8307af8d0b900375914194ff3

mkMod :: Context -> AST -> AST -> IO AST

Create an AST node representing arg1 mod arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga8e350ac77e6b8fe805f57efe196e7713

mkRem :: Context -> AST -> AST -> IO AST

Create an AST node representing arg1 rem arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga2fcdb17f9039bbdaddf8a30d037bd9ff

mkLt :: Context -> AST -> AST -> IO AST

Create less than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga58a3dc67c5de52cf599c346803ba1534

mkLe :: Context -> AST -> AST -> IO AST

Create less than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaa9a33d11096841f4e8c407f1578bc0bf

mkGt :: Context -> AST -> AST -> IO AST

Create greater than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga46167b86067586bb742c0557d7babfd3

mkGe :: Context -> AST -> AST -> IO AST

Create greater than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad9245cbadb80b192323d01a8360fb942

mkInt2Real :: Context -> AST -> IO AST

Coerce an integer to a real.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7130641e614c7ebafd28ae16a7681a21

mkReal2Int :: Context -> AST -> IO AST

Coerce a real to an integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga759b6563ba1204aae55289009a3fdc6d

mkIsInt :: Context -> AST -> IO AST

Check if a real number is an integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaac2ad0fb04e4900fdb4add438d137ad3

Bit-vectors

mkBvnot :: Context -> AST -> IO AST

Bitwise negation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga36cf75c92c54c1ca633a230344f23080

mkBvredand :: Context -> AST -> IO AST

Take conjunction of bits in vector, return vector of length 1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaccc04f2b58903279b1b3be589b00a7d8

mkBvredor :: Context -> AST -> IO AST

Take disjunction of bits in vector, return vector of length 1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafd18e127c0586abf47ad9cd96895f7d2

mkBvand :: Context -> AST -> AST -> IO AST

Bitwise and.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab96e0ea55334cbcd5a0e79323b57615d

mkBvor :: Context -> AST -> AST -> IO AST

Bitwise or.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga77a6ae233fb3371d187c6d559b2843f5

mkBvxor :: Context -> AST -> AST -> IO AST

Bitwise exclusive-or.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga0a3821ea00b1c762205f73e4bc29e7d8

mkBvnand :: Context -> AST -> AST -> IO AST

Bitwise nand.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga96dc37d36efd658fff5b2b4df49b0e61

mkBvnor :: Context -> AST -> AST -> IO AST

Bitwise nor.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gabf15059e9e8a2eafe4929fdfd259aadb

mkBvxnor :: Context -> AST -> AST -> IO AST

Bitwise xnor.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga784f5ca36a4b03b93c67242cc94b21d6

mkBvneg :: Context -> AST -> IO AST

Standard two's complement unary minus.

Reference: <http:research.microsoft.comen-usumredmondprojectsz3group__capi.html#ga0c78be00c03eda4ed6a983224ed5c7b7

mkBvadd :: Context -> AST -> AST -> IO AST

Standard two's complement addition.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga819814e33573f3f9948b32fdc5311158

mkBvsub :: Context -> AST -> AST -> IO AST

Standard two's complement subtraction.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga688c9aa1347888c7a51be4e46c19178e

mkBvmul :: Context -> AST -> AST -> IO AST

Standard two's complement multiplication.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6abd3dde2a1ceff1704cf7221a72258c

mkBvudiv :: Context -> AST -> AST -> IO AST

Unsigned division.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga56ce0cd61666c6f8cf5777286f590544

mkBvsdiv :: Context -> AST -> AST -> IO AST

Two's complement signed division.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad240fedb2fda1c1005b8e9d3c7f3d5a0

mkBvurem :: Context -> AST -> AST -> IO AST

Unsigned remainder.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5df4298ec835e43ddc9e3e0bae690c8d

mkBvsrem :: Context -> AST -> AST -> IO AST

Two's complement signed remainder (sign follows dividend).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga46c18a3042fca174fe659d3185693db1

mkBvsmod :: Context -> AST -> AST -> IO AST

Two's complement signed remainder (sign follows divisor).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga95dac8e6eecb50f63cb82038560e0879

mkBvult :: Context -> AST -> AST -> IO AST

Unsigned less than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5774b22e93abcaf9b594672af6c7c3c4

mkBvslt :: Context -> AST -> AST -> IO AST

Two's complement signed less than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga8ce08af4ed1fbdf08d4d6e63d171663a

mkBvule :: Context -> AST -> AST -> IO AST

Unsigned less than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab738b89de0410e70c089d3ac9e696e87

mkBvsle :: Context -> AST -> AST -> IO AST

Two's complement signed less than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab7c026feb93e7d2eab180e96f1e6255d

mkBvuge :: Context -> AST -> AST -> IO AST

Unsigned greater than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gade58fbfcf61b67bf8c4a441490d3c4df

mkBvsge :: Context -> AST -> AST -> IO AST

Two's complement signed greater than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaeec3414c0e8a90a6aa5a23af36bf6dc5

mkBvugt :: Context -> AST -> AST -> IO AST

Unsigned greater than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga063ab9f16246c99e5c1c893613927ee3

mkBvsgt :: Context -> AST -> AST -> IO AST

Two's complement signed greater than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4e93a985aa2a7812c7c11a2c65d7c5f0

mkConcat :: Context -> AST -> AST -> IO AST

Concatenate the given bit-vectors.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae774128fa5e9ff7458a36bd10e6ca0fa

mkExtract :: Context -> Int -> Int -> AST -> IO AST

Extract the bits high down to low from a bitvector of size m to yield a new bitvector of size n, where n = high - low + 1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga32d2fe7563f3e6b114c1b97b205d4317

mkSignExt :: Context -> Int -> AST -> IO AST

Sign-extend of the given bit-vector to the (signed) equivalent bitvector of size m+i, where m is the size of the given bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad29099270b36d0680bb54b560353c10e

mkZeroExt :: Context -> Int -> AST -> IO AST

Extend the given bit-vector with zeros to the (unsigned) equivalent bitvector of size m+i, where m is the size of the given bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac9322fae11365a78640baf9078c428b3

mkRepeat :: Context -> Int -> AST -> IO AST

Repeat the given bit-vector up length i.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga03e81721502ea225c264d1f556c9119d

mkBvshl :: Context -> AST -> AST -> IO AST

Shift left.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac8d5e776c786c1172fa0d7dfede454e1

mkBvlshr :: Context -> AST -> AST -> IO AST

Logical shift right.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac59645a6edadad79a201f417e4e0c512

mkBvashr :: Context -> AST -> AST -> IO AST

Arithmetic shift right.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga674b580ad605ba1c2c9f9d3748be87c4

mkRotateLeft :: Context -> Int -> AST -> IO AST

Rotate bits of t1 to the left i times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4932b7d08fea079dd903cd857a52dcda

mkRotateRight :: Context -> Int -> AST -> IO AST

Rotate bits of t1 to the right i times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga3b94e1bf87ecd1a1858af8ebc1da4a1c

mkExtRotateLeft :: Context -> AST -> AST -> IO AST

Rotate bits of t1 to the left t2 times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaf46f1cb80e5a56044591a76e7c89e5e7

mkExtRotateRight :: Context -> AST -> AST -> IO AST

Rotate bits of t1 to the right t2 times.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gabb227526c592b523879083f12aab281f

mkInt2bv :: Context -> Int -> AST -> IO AST

Create an n bit bit-vector from the integer argument t1.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga35f89eb05df43fbd9cce7200cc1f30b5

mkBv2int :: Context -> AST -> Bool -> IO AST

Create an integer from the bit-vector argument t1. If is_signed is false, then the bit-vector t1 is treated as unsigned. So the result is non-negative and in the range [0..2^N-1], where N are the number of bits in t1. If is_signed is true, t1 is treated as a signed bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac87b227dc3821d57258d7f53a28323d4

mkBvnegNoOverflow :: Context -> AST -> IO AST

Check that bit-wise negation does not overflow when t1 is interpreted as a signed bit-vector.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae9c5d72605ddcd0e76657341eaccb6c7

mkBvaddNoOverflow :: Context -> AST -> AST -> Bool -> IO AST

Create a predicate that checks that the bit-wise addition of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga88f6b5ec876f05e0d7ba51e96c4b077f

mkBvaddNoUnderflow :: Context -> AST -> AST -> IO AST

Create a predicate that checks that the bit-wise signed addition of t1 and t2 does not underflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1e2b1927cf4e50000c1600d47a152947

mkBvsubNoOverflow :: Context -> AST -> AST -> IO AST

Create a predicate that checks that the bit-wise signed subtraction of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga785f8127b87e0b42130e6d8f52167d7c

mkBvsubNoUnderflow :: Context -> AST -> AST -> IO AST

Create a predicate that checks that the bit-wise subtraction of t1 and t2 does not underflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6480850f9fa01e14aea936c88ff184c4

mkBvmulNoOverflow :: Context -> AST -> AST -> Bool -> IO AST

Create a predicate that checks that the bit-wise multiplication of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga86f4415719d295a2f6845c70b3aaa1df

mkBvmulNoUnderflow :: Context -> AST -> AST -> IO AST

Create a predicate that checks that the bit-wise signed multiplication of t1 and t2 does not underflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga501ccc01d737aad3ede5699741717fda

mkBvsdivNoOverflow :: Context -> AST -> AST -> IO AST

Create a predicate that checks that the bit-wise signed division of t1 and t2 does not overflow.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaa17e7b2c33dfe2abbd74d390927ae83e

Arrays

mkSelect :: Context -> AST -> AST -> IO AST

Array read. The argument a is the array and i is the index of the array that gets read.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga38f423f3683379e7f597a7fe59eccb67

mkStore :: Context -> AST -> AST -> AST -> IO AST

Array update.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae305a4f54b4a64f7e5973ae6ccb13593

mkConstArray :: Context -> Sort -> AST -> IO AST

Create the constant array.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga84ea6f0c32b99c70033feaa8f00e8f2d

mkMap :: Context -> FuncDecl -> Int -> [AST] -> IO AST

map f on the the argument arrays.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga9150242d9430a8c3d55d2ca3b9a4362d

mkArrayDefault :: Context -> AST -> IO AST

Access the array default value. Produces the default range value, for arrays that can be represented as finite maps with a default range value.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga78e89cca82f0ab4d5f4e662e5e5fba7d

Numerals

mkNumeral :: Context -> String -> Sort -> IO AST

Create a numeral of a given sort.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac8aca397e32ca33618d8024bff32948c

mkInt :: Integral a => Context -> a -> IO AST

Create a numeral of sort int.

mkReal :: Real r => Context -> r -> IO AST

Create a numeral of sort real.

Quantifiers

mkPattern :: Context -> [AST] -> IO Pattern

mkBound :: Context -> Int -> Sort -> IO AST

mkForall :: Context -> [Pattern] -> [Symbol] -> [Sort] -> AST -> IO AST

mkExists :: Context -> [Pattern] -> [Symbol] -> [Sort] -> AST -> IO AST

Accessors

getBvSortSize :: Context -> Sort -> IO Int

Return the size of the given bit-vector sort.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga8fc3550edace7bc046e16d1f96ddb419

getSort :: Context -> AST -> IO Sort

Return the sort of an AST node.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga0a4dac7e9397ff067136354cd33cb933

getBool :: Context -> AST -> IO (Maybe Bool)

Return Z3_L_TRUE if a is true, Z3_L_FALSE if it is false, and Z3_L_UNDEF otherwise.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga133aaa1ec31af9b570ed7627a3c8c5a4

getInt :: Context -> AST -> IO Integer

Return Z3Int value

getReal :: Context -> AST -> IO Rational

Return Z3Real value

Models

data FuncModel

The interpretation of a function is a mapping part (arguments to values) and an 'else' part.

Constructors

FuncModel

Fields

interpMap :: [([AST], AST)]

interpElse :: AST

eval :: Context -> Model -> AST -> IO (Maybe AST)

Evaluate an AST node in the given model.

evalT :: Traversable t => Context -> Model -> t AST -> IO (Maybe (t AST))

Evaluate a collection of AST nodes in the given model.

evalFunc :: Context -> Model -> FuncDecl -> IO (Maybe FuncModel)

Evaluate a function declaration to a list of argument/value pairs.

evalArray :: Context -> Model -> AST -> IO (Maybe FuncModel)

Evaluate an array as a function, if possible.

getFuncInterp :: Context -> Model -> FuncDecl -> IO (Maybe FuncInterp)

Return the interpretation of the function f in the model m. Return NULL, if the model does not assign an interpretation for f. That should be interpreted as: the f does not matter.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafb9cc5eca9564d8a849c154c5a4a8633

isAsArray :: Context -> AST -> IO Bool

The (_ as-array f) AST node is a construct for assigning interpretations for arrays in Z3. It is the array such that forall indices i we have that (select (_ as-array f) i) is equal to (f i). This procedure returns Z3_TRUE if the a is an as-array AST node.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4674da67d226bfb16861829b9f129cfa

getAsArrayFuncDecl :: Context -> AST -> IO FuncDecl

Return the function declaration f associated with a (_ as_array f) node.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7d9262dc6e79f2aeb23fd4a383589dda

funcInterpGetNumEntries :: Context -> FuncInterp -> IO Int

Return the number of entries in the given function interpretation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga2bab9ae1444940e7593729beec279844

funcInterpGetEntry :: Context -> FuncInterp -> Int -> IO FuncEntry

Return a _point_ of the given function intepretation. It represents the value of f in a particular point.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaf157e1e1cd8c0cfe6a21be6370f659da

funcInterpGetElse :: Context -> FuncInterp -> IO AST

Return the 'else' value of the given function interpretation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga46de7559826ba71b8488d727cba1fb64

funcInterpGetArity :: Context -> FuncInterp -> IO Int

Return the arity (number of arguments) of the given function interpretation.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaca22cbdb6f7787aaae5d814f2ab383d8

funcEntryGetValue :: Context -> FuncEntry -> IO AST

Return the value of this point.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga9fd65e2ab039aa8e40608c2ecf7084da

funcEntryGetNumArgs :: Context -> FuncEntry -> IO Int

Return the number of arguments in a Z3_func_entry object.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga51aed8c5bc4b1f53f0c371312de3ce1a

funcEntryGetArg :: Context -> FuncEntry -> Int -> IO AST

Return an argument of a Z3_func_entry object.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6fe03fe3c824fceb52766a4d8c2cbeab

modelToString :: Context -> Model -> IO String

Convert the given model into a string.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaf36d49862a8c0d20dd5e6508eef5f8af

showModel :: Context -> Model -> IO String

Alias for modelToString.

Constraints

assertCnstr :: Context -> AST -> IO ()

Assert a constraing into the logical context.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1a05ff73a564ae7256a2257048a4680a

check :: Context -> IO Result

Check whether the given logical context is consistent or not.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga72055cfbae81bd174abed32a83e50b03

getModel :: Context -> IO (Result, Maybe Model)

Get model (logical context is consistent)

Reference : http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaff310fef80ac8a82d0a51417e073ec0a

delModel :: Context -> Model -> IO ()

Delete a model object.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga0cc98d3ce68047f873e119bccaabdbee

push :: Context -> IO ()

Create a backtracking point.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad651ad68c7a060cbb5616349233cb10f

pop :: Context -> Int -> IO ()

Backtrack n backtracking points.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab2b3a542006c86c8d86dc37872f88b61

Parameters

mkParams :: Context -> IO Params

paramsSetBool :: Context -> Params -> Symbol -> Bool -> IO ()

paramsSetUInt :: Context -> Params -> Symbol -> Int -> IO ()

paramsSetDouble :: Context -> Params -> Symbol -> Double -> IO ()

paramsSetSymbol :: Context -> Params -> Symbol -> Symbol -> IO ()

paramsToString :: Context -> Params -> IO String

Solvers

data Logic

Solvers available in Z3.

These are described at http://smtlib.cs.uiowa.edu/logics.html

WARNING: Support for solvers is fragile, you may experience segmentation faults!

Constructors

AUFLIA
Closed formulas over the theory of linear integer arithmetic and arrays extended with free sort and function symbols but restricted to arrays with integer indices and values.

AUFLIRA
Closed linear formulas with free sort and function symbols over one- and two-dimentional arrays of integer index and real value.

AUFNIRA
Closed formulas with free function and predicate symbols over a theory of arrays of arrays of integer index and real value.

LRA
Closed linear formulas in linear real arithmetic.

QF_ABV
Closed quantifier-free formulas over the theory of bitvectors and bitvector arrays.

QF_AUFBV
Closed quantifier-free formulas over the theory of bitvectors and bitvector arrays extended with free sort and function symbols.

QF_AUFLIA
Closed quantifier-free linear formulas over the theory of integer arrays extended with free sort and function symbols.

QF_AX
Closed quantifier-free formulas over the theory of arrays with extensionality.

QF_BV
Closed quantifier-free formulas over the theory of fixed-size bitvectors.

QF_IDL
Difference Logic over the integers. In essence, Boolean combinations of inequations of the form x - y < b where x and y are integer variables and b is an integer constant.

QF_LIA
Unquantified linear integer arithmetic. In essence, Boolean combinations of inequations between linear polynomials over integer variables.

QF_LRA
Unquantified linear real arithmetic. In essence, Boolean combinations of inequations between linear polynomials over real variables.

QF_NIA
Quantifier-free integer arithmetic.

QF_NRA
Quantifier-free real arithmetic.

QF_RDL
Difference Logic over the reals. In essence, Boolean combinations of inequations of the form x - y < b where x and y are real variables and b is a rational constant.

QF_UF
Unquantified formulas built over a signature of uninterpreted (i.e., free) sort and function symbols.

QF_UFBV
Unquantified formulas over bitvectors with uninterpreted sort function and symbols.

QF_UFIDL
Difference Logic over the integers (in essence) but with uninterpreted sort and function symbols.

QF_UFLIA
Unquantified linear integer arithmetic with uninterpreted sort and function symbols.

QF_UFLRA
Unquantified linear real arithmetic with uninterpreted sort and function symbols.

QF_UFNRA
Unquantified non-linear real arithmetic with uninterpreted sort and function symbols.

UFLRA
Linear real arithmetic with uninterpreted sort and function symbols.

UFNIA
Non-linear integer arithmetic with uninterpreted sort and function symbols.

mkSolver :: Context -> IO Solver

mkSimpleSolver :: Context -> IO Solver

mkSolverForLogic :: Context -> Logic -> IO Solver

solverSetParams :: Context -> Solver -> Params -> IO ()

solverPush :: Context -> Solver -> IO ()

solverPop :: Context -> Solver -> Int -> IO ()

solverReset :: Context -> Solver -> IO ()

solverAssertCnstr :: Context -> Solver -> AST -> IO ()

solverAssertAndTrack :: Context -> Solver -> AST -> AST -> IO ()

solverCheck :: Context -> Solver -> IO Result

solverCheckAndGetModel :: Context -> Solver -> IO (Result, Maybe Model)

solverGetReasonUnknown :: Context -> Solver -> IO String

String Conversion

data ASTPrintMode

Pretty-printing mode for converting ASTs to strings. The mode can be one of the following:

Z3_PRINT_SMTLIB_FULL: Print AST nodes in SMTLIB verbose format.

Z3_PRINT_LOW_LEVEL: Print AST nodes using a low-level format.

Z3_PRINT_SMTLIB_COMPLIANT: Print AST nodes in SMTLIB 1.x compliant format.

Z3_PRINT_SMTLIB2_COMPLIANT: Print AST nodes in SMTLIB 2.x compliant format.

Constructors

Z3_PRINT_SMTLIB_FULL

Z3_PRINT_LOW_LEVEL

Z3_PRINT_SMTLIB_COMPLIANT

Z3_PRINT_SMTLIB2_COMPLIANT

setASTPrintMode :: Context -> ASTPrintMode -> IO ()

Set the pretty-printing mode for converting ASTs to strings.

astToString :: Context -> AST -> IO String

Convert an AST to a string.

patternToString :: Context -> Pattern -> IO String

Convert a pattern to a string.

sortToString :: Context -> Sort -> IO String

Convert a sort to a string.

funcDeclToString :: Context -> FuncDecl -> IO String

Convert a FuncDecl to a string.

benchmarkToSMTLibString

Arguments

:: Context

-> String
name

-> String
logic

-> String
status

-> String
attributes

-> [AST]
assumptions

-> AST
formula

-> IO String

Convert the given benchmark into SMT-LIB formatted string.

The output format can be configured via setASTPrintMode.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaf93844a5964ad8dee609fac3470d86e4

Error Handling

data Z3Error

Z3 exceptions.

Constructors

Z3Error

Fields

errCode :: Z3ErrorCode

errMsg :: String

data Z3ErrorCode

Constructors

SortError

IOB

InvalidArg

ParserError

NoParser

InvalidPattern

MemoutFail

FileAccessError

InternalFatal

InvalidUsage

DecRefError

Z3Exception