Z3.Base

Description

Medium-level bindings, highly inspired by yices-painless.

Types

data Config

A Z3 configuration object.

Notes:

The resource is automatically managed by the Haskell garbage collector, and the structure is automatically deleted once it is out of scope (no need to call z3_del_config).

data Context

A Z3 logical context.

Notes:

The resource is automatically managed by the Haskell garbage collector, and the structure is automatically deleted once it is out of scope (no need to call z3_del_context).

data Symbol

A Z3 Lisp-link symbol.

data AST a

A Z3 AST node.

TODO: Does the extra type safety provided by the phantom type worth complicating the higher-level layers such as Monad ?

data Sort a

Kind of Z3 AST representing types.

data FuncDecl a

Kind of AST used to represent function symbols.

data App a

A kind of Z3 AST used to represent constant and function declarations.

data Pattern

A kind of AST used to represent pattern and multi-patterns used to guide quantifier instantiation.

data Model

A model for the constraints asserted into the logical context.

castAST :: forall a b. (Z3Type a, Z3Type b) => AST a -> Maybe (AST b)

Cast an 'AST a' to 'AST b' when a and b are the same type.

This is useful when unpacking an existentially quantified AST.

Satisfiability result

data Result a

Result of a satisfiability check.

Constructors

Sat a

Unsat

Undef

Z3 types

class (Eq a, Show a, Typeable a) => Z3Type a where

A Z3 type

Methods

mkSort :: Context -> IO (Sort a)

mkValue :: Context -> a -> IO (AST a)

getValue :: Context -> AST a -> IO a

class Z3Fun a

A Function type

class (Z3Type a, Num a) => Z3Num a

A Z3 numeric type

Configuration

mkConfig :: IO Config

Create a configuration.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7d6c40d9b79fe8a8851cc8540970787f

setParamValue :: Config -> String -> String -> IO ()

Set a configuration parameter.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga001ade87a1671fe77d7e53ed0f4f1ec3

See: http://research.microsoft.com/en-us/um/redmond/projects/z3/config.html

set_MODEL :: Config -> Bool -> IO ()

Set the MODEL configuration parameter.

default: True, enable/disable model construction.

set_MODEL_PARTIAL :: Config -> Bool -> IO ()

Set the MODEL_PARTIAL configuration parameter.

default: False, enable/disable partial function interpretations.

set_WELL_SORTED_CHECK :: Config -> Bool -> IO ()

Set the WELL_SORTED_CHECK configuration parameter.

default: True, enable/disable type checker.

Context

mkContext :: Config -> IO Context

Create a context using the given configuration.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga0bd93cfab4d749dd3e2f2a4416820a46

Symbols

mkStringSymbol :: Context -> String -> IO Symbol

Create a Z3 symbol using a string.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gafebb0d3c212927cf7834c3a20a84ecae

Sorts

mkBoolSort :: Context -> IO (Sort Bool)

Create the Boolean type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacdc73510b69a010b71793d429015f342

mkIntSort :: Context -> IO (Sort Integer)

Create an integer type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga6cd426ab5748653b77d389fd3eac1015

mkRealSort :: Context -> IO (Sort Rational)

Create a real type.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga40ef93b9738485caed6dc84631c3c1a0

Constants and Applications

mkFuncDecl :: forall t. Z3Fun t => Context -> Symbol -> IO (FuncDecl t)

A Z3 function

mkApp1 :: (Z3Type a, Z3Type b) => Context -> FuncDecl (a -> b) -> AST a -> IO (AST b)

Create a constant or function application.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga33a202d86bf628bfab9b6f437536cebe

mkApp2 :: (Z3Type a, Z3Type b, Z3Type c) => Context -> FuncDecl (a -> b -> c) -> AST a -> AST b -> IO (AST c)

mkApp3 :: (Z3Type a, Z3Type b, Z3Type c, Z3Type d) => Context -> FuncDecl (a -> b -> c -> d) -> AST a -> AST b -> AST c -> IO (AST d)

mkApp4 :: (Z3Type a, Z3Type b, Z3Type c, Z3Type d, Z3Type e) => Context -> FuncDecl (a -> b -> c -> d -> e) -> AST a -> AST b -> AST c -> AST d -> IO (AST e)

mkApp5 :: (Z3Type a, Z3Type b, Z3Type c, Z3Type d, Z3Type e, Z3Type f) => Context -> FuncDecl (a -> b -> c -> d -> e -> f) -> AST a -> AST b -> AST c -> AST d -> AST e -> IO (AST f)

mkConst :: Z3Type a => Context -> Symbol -> Sort a -> IO (AST a)

Declare and create a constant.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga093c9703393f33ae282ec5e8729354ef

mkTrue :: Context -> IO (AST Bool)

Create an AST node representing true.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gae898e7380409bbc57b56cc5205ef1db7

mkFalse :: Context -> IO (AST Bool)

Create an AST node representing false.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga5952ac17671117a02001fed6575c778d

mkEq :: Z3Type a => Context -> AST a -> AST a -> IO (AST Bool)

Create an AST node representing l = r.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga95a19ce675b70e22bb0401f7137af37c

mkNot :: Context -> AST Bool -> IO (AST Bool)

Create an AST node representing not(a).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga3329538091996eb7b3dc677760a61072

mkIte :: Context -> AST Bool -> AST a -> AST a -> IO (AST a)

Create an AST node representing an if-then-else: ite(t1, t2, t3).

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga94417eed5c36e1ad48bcfc8ad6e83547

mkIff :: Context -> AST Bool -> AST Bool -> IO (AST Bool)

Create an AST node representing t1 iff t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga930a8e844d345fbebc498ac43a696042

mkImplies :: Context -> AST Bool -> AST Bool -> IO (AST Bool)

Create an AST node representing t1 implies t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac829c0e25bbbd30343bf073f7b524517

mkXor :: Context -> AST Bool -> AST Bool -> IO (AST Bool)

Create an AST node representing t1 xor t2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacc6d1b848032dec0c4617b594d4229ec

mkAnd :: Context -> [AST Bool] -> IO (AST Bool)

Create an AST node representing args[0] and ... and args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gacde98ce4a8ed1dde50b9669db4838c61

mkOr :: Context -> [AST Bool] -> IO (AST Bool)

Create an AST node representing args[0] or ... or args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga00866d16331d505620a6c515302021f9

mkAdd :: Z3Num a => Context -> [AST a] -> IO (AST a)

Create an AST node representing args[0] + ... + args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4e4ac0a4e53eee0b4b0ef159ed7d0cd5

mkMul :: Z3Num a => Context -> [AST a] -> IO (AST a)

Create an AST node representing args[0] * ... * args[num_args-1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gab9affbf8401a18eea474b59ad4adc890

mkSub :: Z3Num a => Context -> [AST a] -> IO (AST a)

Create an AST node representing args[0] - ... - args[num_args - 1].

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga4f5fea9b683f9e674fd8f14d676cc9a9

mkUnaryMinus :: Z3Num a => Context -> AST a -> IO (AST a)

Create an AST node representing -arg.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gadcd2929ad732937e25f34277ce4988ea

mkDiv :: Z3Num a => Context -> AST a -> AST a -> IO (AST a)

Create an AST node representing arg1 div arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1ac60ee8307af8d0b900375914194ff3

mkMod :: Context -> AST Integer -> AST Integer -> IO (AST Integer)

Create an AST node representing arg1 mod arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga8e350ac77e6b8fe805f57efe196e7713

mkRem :: Context -> AST Integer -> AST Integer -> IO (AST Integer)

Create an AST node representing arg1 rem arg2.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga2fcdb17f9039bbdaddf8a30d037bd9ff

mkLt :: Z3Num a => Context -> AST a -> AST a -> IO (AST Bool)

Create less than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga58a3dc67c5de52cf599c346803ba1534

mkLe :: Z3Num a => Context -> AST a -> AST a -> IO (AST Bool)

Create less than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaa9a33d11096841f4e8c407f1578bc0bf

mkGt :: Z3Num a => Context -> AST a -> AST a -> IO (AST Bool)

Create greater than.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga46167b86067586bb742c0557d7babfd3

mkGe :: Z3Num a => Context -> AST a -> AST a -> IO (AST Bool)

Create greater than or equal to.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gad9245cbadb80b192323d01a8360fb942

mkInt2Real :: Context -> AST Integer -> IO (AST Rational)

Coerce an integer to a real.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga7130641e614c7ebafd28ae16a7681a21

mkReal2Int :: Context -> AST Rational -> IO (AST Integer)

Coerce a real to an integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga759b6563ba1204aae55289009a3fdc6d

mkIsInt :: Context -> AST Rational -> IO (AST Bool)

Check if a real number is an integer.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaac2ad0fb04e4900fdb4add438d137ad3

Numerals

mkNumeral :: Z3Num a => Context -> String -> Sort a -> IO (AST a)

Create a numeral of a given sort.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gac8aca397e32ca33618d8024bff32948c

mkInt :: Integral a => Context -> a -> IO (AST Integer)

Create a numeral of sort int.

mkReal :: Real r => Context -> r -> IO (AST Rational)

Create a numeral of sort real.

Quantifiers

mkPattern :: Context -> [AST a] -> IO Pattern

mkBound :: Context -> Int -> Sort a -> IO (AST a)

mkForall :: Context -> [Pattern] -> Symbol -> Sort a -> AST Bool -> IO (AST Bool)

Accessors

getBool :: Context -> AST Bool -> IO (Maybe Bool)

Return Z3_L_TRUE if a is true, Z3_L_FALSE if it is false, and Z3_L_UNDEF otherwise.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga133aaa1ec31af9b570ed7627a3c8c5a4

getInt :: Context -> AST Integer -> IO Integer

Return Z3Int value

getReal :: Context -> AST Rational -> IO Rational

Return Z3Real value

Models

eval :: Context -> Model -> AST a -> IO (Maybe (AST a))

Evaluate an AST node in the given model.

Constraints

assertCnstr :: Context -> AST Bool -> IO ()

Assert a constraing into the logical context.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga1a05ff73a564ae7256a2257048a4680a

check :: Context -> IO (Result ())

Check whether the given logical context is consistent or not.

Reference: http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#ga72055cfbae81bd174abed32a83e50b03

getModel :: Context -> IO (Result Model)

Get model (logical context is consistent)

Reference : http://research.microsoft.com/en-us/um/redmond/projects/z3/group__capi.html#gaff310fef80ac8a82d0a51417e073ec0a