Z3.Lang.Nat

Description

Documentation

data Nat

This type allows to reason about natural numbers.

Naturals are just integers plus a (>= 0) invariant, and we ensure that this invariant is always preserved by transparently adding new assertions to the context.

Note that arithmetic on naturals must result in natural numbers, otherwise the problem becomes unsatisfiable.