Z3.Lang.Prelude

Description

Z3 script

data Z3 a

Z3 monad.

data Result a

Result of a satisfiability check.

Constructors

Sat a

Unsat

Undef

evalZ3 :: Z3 a -> IO a

Eval a Z3 script.

data Args

Constructors

Args

Fields

softTimeout :: Maybe Int

soft timeout (in milliseconds)

stdArgs :: Args

evalZ3With :: Args -> Z3 a -> IO a

Eval a Z3 script.

Commands

var :: forall a. IsTy a => Z3 (Expr a)

Declare skolem variables.

fun1 :: (IsTy a, IsTy b) => Z3 (Expr a -> Expr b)

Declare uninterpreted function of arity 1.

fun2 :: (IsTy a, IsTy b, IsTy c) => Z3 (Expr a -> Expr b -> Expr c)

Declare uninterpreted function of arity 2.

fun3 :: (IsTy a, IsTy b, IsTy c, IsTy d) => Z3 (Expr a -> Expr b -> Expr c -> Expr d)

Declare uninterpreted function of arity 3.

fun4 :: (IsTy a, IsTy b, IsTy c, IsTy d, IsTy e) => Z3 (Expr a -> Expr b -> Expr c -> Expr d -> Expr e)

Declare uninterpreted function of arity 4.

fun5 :: (IsTy a, IsTy b, IsTy c, IsTy d, IsTy e, IsTy f) => Z3 (Expr a -> Expr b -> Expr c -> Expr d -> Expr e -> Expr f)

Declare uninterpreted function of arity 5.

assert :: Expr Bool -> Z3 ()

Make assertion in current context.

let_ :: IsTy a => Expr a -> Z3 (Expr a)

Introduce an auxiliary declaration to name a given expression.

If you really want sharing use this instead of Haskell's let.

check :: Z3 (Result ())

Check satisfiability.

checkModel :: forall a. IsTy a => Expr a -> Z3 (Result a)

Check satisfiability and evaluate the given expression if a model exists.

Expressions

data Expr

Abstract syntax.

data Pattern where

Quantifier pattern.

Constructors

Pat :: IsTy a => Expr a -> Pattern

class (Eq a, Show a, Typeable a, Compilable (Expr a)) => IsTy a

Types for expressions.

class (IsTy a, Num a, Z3Num (TypeZ3 a)) => IsNum a

Numeric types.

class (IsNum a, Integral a, TypeZ3 a ~ Integer) => IsInt a

Typeclass for Haskell Z3 numbers of int sort in Z3.

class (IsNum a, Fractional a, Real a, TypeZ3 a ~ Rational) => IsReal a

Typeclass for Haskell Z3 numbers of real sort in Z3.

literal :: IsTy a => a -> Expr a

literal constructor.

true, false :: Expr Bool

Boolean literals.

not_ :: Expr Bool -> Expr Bool

Boolean negation

and_ :: [Expr Bool] -> Expr Bool

Boolean variadic and.

(&&*) :: Expr Bool -> Expr Bool -> Expr Bool

Boolean binary and.

or_ :: [Expr Bool] -> Expr Bool

Boolean variadic or.

(||*) :: Expr Bool -> Expr Bool -> Expr Bool

Boolean binary or.

xor :: Expr Bool -> Expr Bool -> Expr Bool

Boolean binary xor.

implies :: Expr Bool -> Expr Bool -> Expr Bool

Boolean implication

(==>) :: Expr Bool -> Expr Bool -> Expr Bool

An alias for implies.

iff :: Expr Bool -> Expr Bool -> Expr Bool

Boolean if and only if.

(<=>) :: Expr Bool -> Expr Bool -> Expr Bool

An alias for iff.

forall :: forall a. IsTy a => (Expr a -> Expr Bool) -> Expr Bool

Forall formula.

forallP

Arguments

:: IsTy a

=> (Expr a -> Expr Bool)
body

-> (Expr a -> Pattern)
pattern

-> Expr Bool

(//) :: IsInt a => Expr a -> Expr a -> Expr a

Integer division.

(%*) :: IsInt a => Expr a -> Expr a -> Expr a

Integer modulo.

(%%) :: IsInt a => Expr a -> Expr a -> Expr a

Integer remainder.

divides :: IsInt a => Expr a -> Expr a -> Expr Bool

k divides n == n %* k ==* 0

(==*) :: IsTy a => Expr a -> Expr a -> Expr Bool

Equals.

(/=*) :: IsTy a => Expr a -> Expr a -> Expr Bool

Not equals.

(<=*) :: IsNum a => Expr a -> Expr a -> Expr Bool

Less or equals than.

(<*) :: IsNum a => Expr a -> Expr a -> Expr Bool

Less than.

(>=*) :: IsNum a => Expr a -> Expr a -> Expr Bool

Greater or equals than.

(>*) :: IsNum a => Expr a -> Expr a -> Expr Bool

Greater than.

min_ :: IsNum a => Expr a -> Expr a -> Expr a

Minimum.

max_ :: IsNum a => Expr a -> Expr a -> Expr a

Maximum.

ite :: IsTy a => Expr Bool -> Expr a -> Expr a -> Expr a

if-then-else.